Extract waving-Dax fit logic into WavingDaxController
