+++ app/src/main/java/com/duckduckgo/app/cta/ui/Cta.kt
@@ -24,7 +24,6 @@ import android.animation.ValueAnimator
 import android.content.Context
 import android.content.res.ColorStateList
 import android.content.res.Configuration
-import android.graphics.RectF
 import android.graphics.drawable.GradientDrawable
 import android.net.Uri
 import android.view.ContextThemeWrapper
@@ -1613,8 +1612,6 @@ sealed class DaxBubbleCta(
             private const val DIALOG_CONTENT_FADE_IN_DURATION = 200L
             private const val HEADER_IMAGE_FADE_IN_DURATION = 300L
             private const val ARROW_DEPTH_ANIMATION_DURATION = 200L
-            private const val FIT_SETTLE_DELAY_MS = 100L
-            private const val DAX_WAVE_START_FRAME = 17
             private const val TYPING_DELAY_MS = 20L
             private const val TYPING_POST_DELAY_MS = 20L
             private const val DISMISS_BORDER_WIDTH_DP = 1.5f
@@ -1639,9 +1636,14 @@ sealed class DaxBubbleCta(
         private var contentFadeInAnimator: AnimatorSet? = null
         private var fadeOutAnimator: AnimatorSet? = null
         private var arrowDepthAnimator: ValueAnimator? = null
-        private var fitArrowAnimator: ValueAnimator? = null
-        private var lastDaxFits: Boolean? = null
-        private val fitRunnable = Runnable { decideFit() }
+
+        private val wavingDaxController: WavingDaxController? by lazy {
+            if (onboardingImprovementsEnabled && this is ShowsWavingDax) {
+                WavingDaxController(showArrow, deviceInfo)
+            } else {
+                null
+            }
+        }
 
         protected fun resolveOnboardingContext(context: Context): Context {
             val themeRes = if (isLightTheme) {
@@ -1692,134 +1694,19 @@ sealed class DaxBubbleCta(
         }
 
         /**
-         * Pure fit test. The Dax fits unless its head intrudes into the card body (above the
-         * top-of-fin line) or its rect overlaps the fin's own region. The Dax may sit in the
-         * fin's vertical band as long as it is horizontally clear of the fin.
-         */
-        internal fun daxFits(
-            daxTop: Int,
-            daxLeft: Int,
-            daxRight: Int,
-            cardBodyBottom: Int,
-            finBottom: Int,
-            finLeft: Int,
-            finRight: Int,
-        ): Boolean {
-            if (daxTop < cardBodyBottom) return false
-            val overlapsFinHorizontally = daxRight > finLeft && daxLeft < finRight
-            if (daxTop < finBottom && overlapsFinHorizontally) return false
-            return true
-        }
-
-        private fun viewBoundsOnScreen(view: View, out: RectF) {
-            out.set(0f, 0f, view.width.toFloat(), view.height.toFloat())
-            view.matrix.mapRect(out)
-            val loc = IntArray(2)
-            (view.parent as? View)?.getLocationOnScreen(loc)
-            out.offset((loc[0] + view.left).toFloat(), (loc[1] + view.top).toFloat())
-        }
-
-        /**
-         * The fin depth [showCta] should apply, or null when [applyFit] owns the fin (waving-Dax
-         * CTAs in portrait/tablet). The null case mirrors [computeDaxFits]'s ownership guard, so
-         * [showCta] and [applyFit] can never both write the fin.
+         * The fin depth [showCta] should apply, or null when the controller owns the fin (waving-Dax
+         * CTAs in portrait/tablet). The null case mirrors the controller's portrait/tablet ownership
+         * guard, so [showCta] and the controller can never both write the fin.
          */
         private fun showCtaFinTarget(container: View): Float? {
             if (onboardingImprovementsEnabled && this is ShowsWavingDax && !container.isPhoneLandscape()) return null
             return if (showArrow && !container.isPhoneLandscape()) 1f else 0f
         }
 
-        /**
-         * Hide eagerly, reveal lazily. A hide (does not fit) applies immediately so the Dax never
-         * lingers or drifts as the keyboard rises; a reveal (fits) is debounced to
-         * [FIT_SETTLE_DELAY_MS] so transient measurements during a content transition or keyboard
-         * settle never flash the Dax/fin in then straight back out. No-op for CTAs without a
-         * waving Dax, in landscape, or when not shown.
-         */
         fun applyFit() {
-            if (!onboardingImprovementsEnabled) return
             val container = ctaView ?: return
-            container.removeCallbacks(fitRunnable)
-            val fits = computeDaxFits() ?: return
-            if (fits) {
-                container.postDelayed(fitRunnable, FIT_SETTLE_DELAY_MS)
-            } else {
-                applyFitResult(false)
-            }
+            wavingDaxController?.applyFit(container)
         }
-
-        private fun decideFit() {
-            if (computeDaxFits() == true) applyFitResult(true)
-        }
-
-        private fun computeDaxFits(): Boolean? {
-            if (this !is ShowsWavingDax) return null
-            val container = ctaView ?: return null
-            if (!container.isShown || container.isPhoneLandscape()) return null
-
-            val dax = container.findViewById<LottieAnimationView>(R.id.wavingDax) ?: return null
-            val cardView = container.findViewById<DaxOnboardingBubbleBrandDesignUpdateCardView>(R.id.brandDesignCardView) ?: return null
-            val finBounds = cardView.arrowBounds() ?: return null
-
-            val daxRect = RectF()
-            viewBoundsOnScreen(dax, daxRect)
-
-            val cardLoc = IntArray(2)
-            cardView.getLocationOnScreen(cardLoc)
-
-            return daxFits(
-                daxTop = daxRect.top.toInt(),
-                daxLeft = daxRect.left.toInt(),
-                daxRight = daxRect.right.toInt(),
-                cardBodyBottom = cardLoc[1] + finBounds.top.toInt(),
-                finBottom = cardLoc[1] + finBounds.bottom.toInt(),
-                finLeft = cardLoc[0] + finBounds.left.toInt(),
-                finRight = cardLoc[0] + finBounds.right.toInt(),
-            )
-        }
-
-        private fun applyFitResult(fits: Boolean) {
-            if (fits == lastDaxFits) return
-            lastDaxFits = fits
-
-            val container = ctaView ?: return
-            val dax = container.findViewById<LottieAnimationView>(R.id.wavingDax) ?: return
-            val cardView = container.findViewById<DaxOnboardingBubbleBrandDesignUpdateCardView>(R.id.brandDesignCardView) ?: return
-
-            if (fits) {
-                if (!dax.isVisible) {
-                    dax.setMinFrame(DAX_WAVE_START_FRAME)
-                    dax.progress = 0f
-                    dax.isVisible = true
-                    dax.post { dax.playAnimation() }
-                }
-            } else {
-                dax.isInvisible = true
-            }
-
-            val finTarget = if (fits && showArrow) 1f else 0f
-            val current = cardView.arrowDepthFraction
-            arrowDepthAnimator?.removeAllUpdateListeners()
-            arrowDepthAnimator?.cancel()
-            fitArrowAnimator?.removeAllUpdateListeners()
-            fitArrowAnimator?.cancel()
-            if (current != finTarget) {
-                fitArrowAnimator = buildArrowDepthAnimator(cardView, current, finTarget).apply { start() }
-            } else {
-                cardView.setArrowDepthFraction(finTarget)
-            }
-        }
-
-        private fun buildArrowDepthAnimator(
-            cardView: DaxOnboardingBubbleBrandDesignUpdateCardView,
-            from: Float,
-            to: Float,
-        ): ValueAnimator =
-            ValueAnimator.ofFloat(from, to).apply {
-                duration = ARROW_DEPTH_ANIMATION_DURATION
-                interpolator = FastOutSlowInInterpolator()
-                addUpdateListener { cardView.setArrowDepthFraction(it.animatedValue as Float) }
-            }
 
         private fun resetAllIncludesExcept(view: View, active: View) {
             getAllContentIncludes(view).forEach { include ->
@@ -1839,7 +1726,7 @@ sealed class DaxBubbleCta(
             ctaView = container
 
             cancelRunningAnimations()
-            lastDaxFits = null
+            wavingDaxController?.reset()
             val isContentTransition = container.alpha > 0f && container.isVisible // card already visible from previous CTA
 
             val cardView = container.findViewById<DaxOnboardingBubbleBrandDesignUpdateCardView>(R.id.brandDesignCardView)
@@ -1900,11 +1787,13 @@ sealed class DaxBubbleCta(
                         val targetDepth = showCtaFinTarget(container)
                         if (targetDepth != null) {
                             val currentDepth = cardView.arrowDepthFraction
-                            if (targetDepth > currentDepth) {
-                                arrowDepthAnimator = buildArrowDepthAnimator(cardView, currentDepth, targetDepth)
+                            if (targetDepth != currentDepth) {
+                                arrowDepthAnimator = ValueAnimator.ofFloat(currentDepth, targetDepth).apply {
+                                    duration = ARROW_DEPTH_ANIMATION_DURATION
+                                    interpolator = FastOutSlowInInterpolator()
+                                    addUpdateListener { cardView.setArrowDepthFraction(it.animatedValue as Float) }
+                                }
                                 animators.add(arrowDepthAnimator!!)
-                            } else if (targetDepth != currentDepth) {
-                                cardView.setArrowDepthFraction(targetDepth)
                             }
                         }
                         contentFadeInAnimator = AnimatorSet().apply {
@@ -2053,11 +1942,7 @@ sealed class DaxBubbleCta(
             arrowDepthAnimator?.removeAllUpdateListeners()
             arrowDepthAnimator?.cancel()
             arrowDepthAnimator = null
-            fitArrowAnimator?.removeAllUpdateListeners()
-            fitArrowAnimator?.cancel()
-            fitArrowAnimator = null
-            lastDaxFits = null
-            ctaView?.removeCallbacks(fitRunnable)
+            ctaView?.let { wavingDaxController?.cancel(it) }
             ctaView?.animate()?.cancel()
         }
